requirements.txt
requests==2.27.1
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0